+++ components/edit/EditShell/CommandBar.tsx
@@ -43,7 +43,7 @@ export function CommandBar({ title, history, insertItems, commands, trailing }: 
   const router = useRouter();
 
   return (
-    <header className="flex h-14 shrink-0 items-center gap-3 border-b border-zinc-200/60 px-5 dark:border-zinc-800/60">
+    <header className="flex h-20 shrink-0 items-center gap-3 border-b border-zinc-200/60 px-8 dark:border-zinc-800/60">
       <div className="flex min-w-0 flex-[2] items-center gap-2">
         {/* Back-to-home — mirrors playback Header's leftmost button so the
             user has the same global-out affordance across modes. */}
