+++ components/edit/SlideNavRail/SlideNavRail.tsx
@@ -1,6 +1,7 @@
 'use client';
 
 import { Fragment, useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { useRouter } from 'next/navigation';
 import { AnimatePresence, Reorder, motion, useReducedMotion } from 'motion/react';
 import { PanelLeftClose, PanelLeftOpen, PlusCircle } from 'lucide-react';
 import { toast } from 'sonner';
@@ -36,6 +37,7 @@ const RAIL_MAX_PX = 360;
  */
 export function SlideNavRail() {
   const { t } = useI18n();
+  const router = useRouter();
   const scenes = useStageStore.use.scenes();
   const currentSceneId = useStageStore.use.currentSceneId();
   const setCurrentSceneId = useStageStore.use.setCurrentSceneId();
@@ -270,19 +272,26 @@ export function SlideNavRail() {
           <div className="absolute right-0.5 top-1/2 -translate-y-1/2 w-0.5 h-8 rounded-full bg-gray-300 dark:bg-gray-600 group-hover:bg-violet-400 dark:group-hover:bg-violet-500 transition-colors" />
         </div>
       )}
-      {/* Header band — height matches playback's logo header (h-10 +
-          mt-3 mb-1 — together ~56px) so when Pro mode opens the chrome
-          height stays consistent with what was here in playback. */}
+      {/* Header band — mirrors playback `SceneSidebar`: OpenMAIC logo
+          on the left (click → home), action cluster on the right.
+          Height (h-10 + mt-3 mb-1 = ~56px) matches playback so the
+          chrome top edge stays at the same screen pixel across the
+          mode swap. */}
       <div
         className={cn(
-          'shrink-0 px-3 mt-3 mb-1',
+          'shrink-0 px-3 mt-3 mb-1 h-10',
           collapsed ? 'flex flex-col items-center gap-1' : 'flex items-center justify-between',
         )}
       >
         {!collapsed && (
-          <span className="text-[10px] uppercase tracking-[0.18em] font-bold text-gray-400 dark:text-gray-500">
-            {t('edit.nav.deckLabel')}
-          </span>
+          <button
+            type="button"
+            onClick={() => router.push('/')}
+            title={t('generation.backToHome')}
+            className="flex items-center gap-2 cursor-pointer rounded-lg px-1.5 -mx-1.5 py-1 -my-1 hover:bg-gray-100/80 dark:hover:bg-gray-800/60 active:scale-[0.97] transition-all duration-150"
+          >
+            <img src="/logo-horizontal.png" alt="OpenMAIC" className="h-6" />
+          </button>
         )}
         <div className={cn('flex items-center gap-1', collapsed && 'flex-col')}>
           <button
